// bam-trivia-project-angular/src/app/models/userquiz.ts
//  public int UserQuizId { get; set; }
//         public int UserId { get; set; }
//         public string Username { get; set; }
//         public int QuizId { get; set; }
//         public int QuizMaxScore { get; set; }
//         public int QuizActualScore { get; set; }
//         public DateTime QuizDate { get; set; }

export class UserQuiz {
    userQuizId: number;
    userId: number;
    userName: string;
    quizId: number;
    quizMaxScore: number;
    quizActualScore: number;
    quizDate: string;
}